+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBDatabase.java
@@ -766,6 +766,15 @@ public class FDBDatabase {
         return transaction;
     }
 
+    /**
+     * Create an {@link FDBDatabaseRunner} for use against this database.
+     * @param contextConfigBuilder options for contexts opened by the new runner
+     * @return a new runner
+     */
+    @Nonnull
+    public FDBDatabaseRunner newRunner(@Nonnull FDBRecordContextConfig.Builder contextConfigBuilder) {
+        return new FDBDatabaseRunnerImpl(this, contextConfigBuilder);
+    }
 
     /**
      * Create an {@link FDBDatabaseRunner} for use against this database.
@@ -773,7 +782,7 @@ public class FDBDatabase {
      */
     @Nonnull
     public FDBDatabaseRunner newRunner() {
-        return new FDBDatabaseRunnerImpl(this);
+        return newRunner(FDBRecordContextConfig.newBuilder().setTransactionTimeoutMillis(factory.getTransactionTimeoutMillis()));
     }
 
     /**
@@ -784,7 +793,9 @@ public class FDBDatabase {
      */
     @Nonnull
     public FDBDatabaseRunner newRunner(@Nullable FDBStoreTimer timer, @Nullable Map<String, String> mdcContext) {
-        return new FDBDatabaseRunnerImpl(this, timer, mdcContext);
+        FDBDatabaseRunner runner = newRunner();
+        runner.getContextConfigBuilder().setTimer(timer).setMdcContext(mdcContext);
+        return runner;
     }
 
     /**
@@ -797,7 +808,9 @@ public class FDBDatabase {
     @Nonnull
     public FDBDatabaseRunner newRunner(@Nullable FDBStoreTimer timer, @Nullable Map<String, String> mdcContext,
                                        @Nullable WeakReadSemantics weakReadSemantics) {
-        return new FDBDatabaseRunnerImpl(this, timer, mdcContext, weakReadSemantics);
+        FDBDatabaseRunner runner = newRunner();
+        runner.getContextConfigBuilder().setTimer(timer).setMdcContext(mdcContext).setWeakReadSemantics(weakReadSemantics);
+        return runner;
     }
 
     /**